Add files via upload
negative input problem is fixed... turns out num had to be just an int, since the user input can be a negative number. and then once in the factorial function is now long long unsigned int to get the largest value and since we already validated the input as a positice integer.